requirements.txt
accelerate==1.6.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
distinctipy==1.3.4
faiss-cpu==1.10.0
ftfy==6.3.1
huggingface-hub==0.29.3
matplotlib==3.9.4
numpy==1.26.4
opencv-contrib-python==4.5.5.64
openpyxl==3.1.5
pandas==2.2.3
peft==0.15.2
pillow==11.1.0
psutil==5.9.8
pytorch-metric-learning==2.8.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.49.0
umap-learn==0.5.7
